+++ libraries/Fake.BKAL/include/FakeSolidMaker.h
@@ -10,8 +10,8 @@ namespace Fake{
     class SolidMaker : public BKAL::ISolidMaker_<Solid>
     {
         public:
-            pIBox makeBox(uint L, uint W, uint H) const override;
-            pICylinder makeCylinder(uint R, uint H) const override;
+            pIBox makeBox(double L, double W, double H) const override;
+            pICylinder makeCylinder(double R, double H) const override;
             pISolid makeFusion(const Solid& base, const Solid& tool) const override;
 
             pISolid fuseBoxAndCylinderFlush() const;
